Move selectable list item to its own file
